refactor: reorganização da estrutura do projeto por feature
Reorganizados os pacotes do projeto para adotar a arquitetura modular por feature.
Cada funcionalidade agora possui seus próprios pacotes Controller, Entity, Repository e Service.
Melhora a legibilidade e facilita a manutenção e escalabilidade do projeto.
