+++ src/main/java/br/com/filipecode/DeskhelpApi/chamado/dto/AtualizarChamadoDTO.java
@@ -1,0 +1,6 @@
+package br.com.filipecode.DeskhelpApi.chamado.dto;
+
+import java.util.UUID;
+
+public record AtualizarChamadoDTO(String status, String prioridade, UUID tecnicoID) {
+}
