+++ src/main/java/br/com/filipecode/DeskhelpApi/chamado/dto/EventoChamadoDTO.java
@@ -1,0 +1,6 @@
+package br.com.filipecode.DeskhelpApi.chamado.dto;
+
+import java.time.LocalDateTime;
+
+public record EventoChamadoDTO(String status, String descricaoEvento, LocalDateTime dataEvento) {
+}
