--- src/main/java/br/com/filipecode/DeskhelpApi/model/dtos/AtualizarChamadoDTO.java
@@ -1,6 +1,0 @@
-package br.com.filipecode.DeskhelpApi.model.dtos;
-
-import java.util.UUID;
-
-public record AtualizarChamadoDTO(String status, String prioridade, UUID tecnicoID) {
-}
