--- src/main/java/br/com/filipecode/DeskhelpApi/model/dtos/ChamadoDTO.java
@@ -1,13 +1,0 @@
-package br.com.filipecode.DeskhelpApi.model.dtos;
-
-import br.com.filipecode.DeskhelpApi.model.enums.Prioridade;
-
-import java.util.UUID;
-
-public record ChamadoDTO(
-        String titulo,
-        String descricao,
-        Prioridade prioridade,
-        UUID usuarioId
-) {
-}
