--- src/main/java/br/com/filipecode/DeskhelpApi/model/dtos/EventoChamadoDTO.java
@@ -1,6 +1,0 @@
-package br.com.filipecode.DeskhelpApi.model.dtos;
-
-import java.time.LocalDateTime;
-
-public record EventoChamadoDTO(String status, String descricaoEvento, LocalDateTime dataEvento) {
-}
