--- src/main/java/br/com/filipecode/DeskhelpApi/model/dtos/TecnicoDTO.java
@@ -1,6 +1,0 @@
-package br.com.filipecode.DeskhelpApi.model.dtos;
-
-public record TecnicoDTO(String nome,
-                         String email,
-                         String especializacao) {
-}
