--- src/main/java/br/com/filipecode/DeskhelpApi/model/dtos/TecnicoRespostaDTO.java
@@ -1,9 +1,0 @@
-package br.com.filipecode.DeskhelpApi.model.dtos;
-
-import java.util.UUID;
-
-public record TecnicoRespostaDTO(UUID id,
-                                 String nome,
-                                 String email,
-                                 String especializacao) {
-}
