--- src/main/java/br/com/filipecode/DeskhelpApi/model/dtos/UsuarioDTO.java
@@ -1,7 +1,0 @@
-package br.com.filipecode.DeskhelpApi.model.dtos;
-
-public record UsuarioDTO(String nome,
-                         String email,
-                         String departamento,
-                         String cargo) {
-}
