--- src/main/java/br/com/filipecode/DeskhelpApi/model/dtos/UsuarioRespostaDTO.java
@@ -1,10 +1,0 @@
-package br.com.filipecode.DeskhelpApi.model.dtos;
-
-import java.util.UUID;
-
-public record UsuarioRespostaDTO(UUID id,
-                                 String nome,
-                                 String email,
-                                 String departamento,
-                                 String cargo) {
-}
