--- src/main/java/br/com/filipecode/DeskhelpApi/model/entities/Chamado.java
@@ -1,52 +1,0 @@
-package br.com.filipecode.DeskhelpApi.model.entities;
-
-import br.com.filipecode.DeskhelpApi.model.enums.Prioridade;
-import br.com.filipecode.DeskhelpApi.model.enums.Status;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-@Entity
-@Table(name = "chamados")
-@Getter
-@Setter
-public class Chamado {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-
-    @Column(nullable = false)
-    private String titulo;
-
-    @Column(nullable = false, columnDefinition = "TEXT")
-    private String descricao;
-
-    @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private Status status;
-
-    @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private Prioridade prioridade;
-
-    @Column(name = "data_criacao", nullable = false)
-    private LocalDateTime dataCriacao;
-
-    @Column(name = "data_atualizacao")
-    private LocalDateTime dataAtualizacao;
-
-    @Column(name = "data_fechamento")
-    private LocalDateTime dataFechamento;
-
-    @ManyToOne
-    @JoinColumn(name = "usuario_id", nullable = false)
-    private Usuario usuario;
-
-    @ManyToOne
-    @JoinColumn(name = "tecnico_id")
-    private Tecnico tecnico;
-}
