--- src/main/java/br/com/filipecode/DeskhelpApi/repositories/AuditoriaRepository.java
@@ -1,9 +1,0 @@
-package br.com.filipecode.DeskhelpApi.repositories;
-
-import br.com.filipecode.DeskhelpApi.model.entities.Auditoria;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.UUID;
-
-public interface AuditoriaRepository extends JpaRepository<Auditoria, UUID> {
-}
