+++ src/main/java/br/com/filipecode/DeskhelpApi/shared/enums/Prioridade.java
@@ -1,0 +1,7 @@
+package br.com.filipecode.DeskhelpApi.shared.enums;
+
+public enum Prioridade {
+    BAIXA,
+    MEDIA,
+    ALTA
+}
