+++ src/main/java/br/com/filipecode/DeskhelpApi/shared/enums/Status.java
@@ -1,0 +1,7 @@
+package br.com.filipecode.DeskhelpApi.shared.enums;
+
+public enum Status {
+    ABERTO,
+    EM_ANDAMENTO,
+    CONCLUIDO
+}
