+++ src/main/java/br/com/filipecode/DeskhelpApi/tecnico/dto/TecnicoDTO.java
@@ -1,0 +1,6 @@
+package br.com.filipecode.DeskhelpApi.tecnico.dto;
+
+public record TecnicoDTO(String nome,
+                         String email,
+                         String especializacao) {
+}
