+++ src/main/java/br/com/filipecode/DeskhelpApi/tecnico/dto/TecnicoRespostaDTO.java
@@ -1,0 +1,9 @@
+package br.com.filipecode.DeskhelpApi.tecnico.dto;
+
+import java.util.UUID;
+
+public record TecnicoRespostaDTO(UUID id,
+                                 String nome,
+                                 String email,
+                                 String especializacao) {
+}
