+++ src/main/java/br/com/filipecode/DeskhelpApi/tecnico/entity/Tecnico.java
@@ -1,0 +1,28 @@
+package br.com.filipecode.DeskhelpApi.tecnico.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+
+import java.util.UUID;
+
+@Entity
+@Table(name = "tecnicos")
+@Getter
+@Setter
+public class Tecnico {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
+
+    @Column(nullable = false)
+    private String nome;
+
+    @Column(nullable = false, unique = true)
+    private String email;
+
+    @Column(nullable = false)
+    private String especializacao;
+}
