+++ src/main/java/br/com/filipecode/DeskhelpApi/usuario/dto/UsuarioDTO.java
@@ -1,0 +1,7 @@
+package br.com.filipecode.DeskhelpApi.usuario.dto;
+
+public record UsuarioDTO(String nome,
+                         String email,
+                         String departamento,
+                         String cargo) {
+}
