+++ src/main/java/br/com/filipecode/DeskhelpApi/usuario/dto/UsuarioRespostaDTO.java
@@ -1,0 +1,10 @@
+package br.com.filipecode.DeskhelpApi.usuario.dto;
+
+import java.util.UUID;
+
+public record UsuarioRespostaDTO(UUID id,
+                                 String nome,
+                                 String email,
+                                 String departamento,
+                                 String cargo) {
+}
